RPG持续构思中，先加个 舒尔特方格 解解闷。

+++ AGameServer/AWebServices/Src/Avatar/AGameObj.cs
@@ -48,7 +48,7 @@ public abstract class AGameObj
             paramComponent.OnSetParamValue(paramname, paramvalue);
         }
     }
-    public void OnSetParamValue(string paramname, float paramvalue)
+    public void OnSetParamValue(string paramname, double paramvalue)
     {
         var paramComponent = OnGetComponent<AComponentParam>();
         if (paramComponent != null)
